+++ packages/office-ui-fabric-react/src/components/Divider/VerticalDivider.classNames.ts
