+++ packages/office-ui-fabric-react/src/components/Divider/index.ts
@@ -1,2 +1,2 @@
-export * from './Divider';
-export * from './Divider.Props';
+export * from './VerticalDivider';
+export * from './VerticalDivider.Props';
